Update the debug handlers options for verbosity
Since the commands are using the standard -v verbosity in the argparser, the help for the command shows that there are four verbosity  levels. At the moment if you run 

python manage.py send_mail -v3 

for instance, it fails with a cryptic message about the key "3" not being present in the LOGGING_LEVEL dictionary. This just duplicates the "2" key so that both 2 and 3 verbosity have the DEBUG message level. An alternative would be to specify the -v option in each of the commands to only allow -v[0/1/2].