Fixes #68 -- improve performance of data migration (#69)
* :zap: Fixes #68 -- improve performance of data migration

Tested against a 13GB SQL dump of the email table (~23K records),
within a cgroup with 300M memory limit. The migration completed
in just under 60s.

* Use iterator to avoid loading entire table in memory (for the queryset cache)
* Defer the message content, as it may contain attachments causing excessive
  memory usage - it is not used in the migration anyway.
* Replace loop over log table with 3 separate SQL update queries

* :bug: Do not prevent migration rollbacks

Discussed in the context of #68 - the reverse data migration is not
implemented (deliberately), but that shouldn't prevent people from
rolling back their migrations in dev-environments and while
debugging/diagnosing issues, for example.